power: pm8921-bms: fix calibration and UVLO issues
The bms is not running a hkadc calibration before calculating the
first SOC. This causes accuracy issues with the power on SOC.

Also when the battery gets in the low voltage range, don't limit
the corrections. We have seen UVLO where even though the BMS was
calculating SOC quickly it was not lowering the SOC fast enough.
Skip limiting ocv change when battery is near cutoff voltage.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Change-Id: I746a9b003135d51d51c7d039d411689ca69571dc