Unhide DEBUG_BUGVERBOSE when EXPERT=y, even if DEBUG_KERNEL=n
CONFIG_DEBUG_BUGVERBOSE defaults to y.  Embedded systems might want to
disable this to save space; however, disabling it requires enabling
CONFIG_DEBUG_KERNEL to see it.  Unhide it for CONFIG_EXPERT=y as well,
so it shows up as an option for embedded users without having to turn on
CONFIG_DEBUG_KERNEL.

Also make the option show up if CONFIG_DEBUG_KERNEL=y and
CONFIG_EXPERT=n, which seems sensible for a debug feature.

Change-Id: I7ba560f4e0ec0d8911e0c83d8e05619b67600928
Signed-off-by: Josh Triplett <josh@joshtriplett.org>