jiffies conversions: Use compile time constants when possible
Date	Fri, 04 Jan 2013 13:15:43 -0800

Do the multiplications and divisions at compile time
instead of runtime when the converted value is a constant.

Make the calculation functions static __always_inline to jiffies.h.

Add #defines with __builtin_constant_p to test and use the
static inline or the runtime functions as appropriate.

Prefix the old exported symbols/functions with __

Change-Id: Ie074c4bed5b0fa268302d77bd480ff6f60e9e491
Signed-off-by: Paul Reioux <reioux@gmail.com>